Development to main (#38)
* updating the custom elements

* packages update from severe out of datenesss

* code cleanup

* renamed to not mess with working files

* access changes

* arrows to functions for hoisting and clarity

* encoding by default

* allow for decoding to crash

* Update logout.html

* changes for idToken and profile return

fixes #133
fixes #134

* Change to idToken

TPEN_USER is tidied.
scope is limited.
required data and access is reduced.
add user profile to the page

* forgot to honor `_id`

* adding profile to user flow

to allow for displayName after signup... we can take it out later if we want to call for more info straight from TPEN

* revenge of the  Oxford comma

* install updates

* Update package.json

* remove .vscode from pr

* checking the signed token

* thin out token handling

* UI shows but assignRole fails

* working except for role removal

* Update manage-api.mjs

* disabled to prevent actions

The assignRole works, but removeRole does not. This prevents the Admins from making confusing data until it is repaired

* Id token internals (#36)

* updating the custom elements

* packages update from severe out of datenesss

* code cleanup

* renamed to not mess with working files

* access changes

* arrows to functions for hoisting and clarity

* encoding by default

* allow for decoding to crash

* Update logout.html

* changes for idToken and profile return

fixes #133
fixes #134

* Change to idToken

TPEN_USER is tidied.
scope is limited.
required data and access is reduced.
add user profile to the page

* forgot to honor `_id`

* adding profile to user flow

to allow for displayName after signup... we can take it out later if we want to call for more info straight from TPEN

* revenge of the  Oxford comma

* install updates

* Update package.json

* remove .vscode from pr

* checking the signed token

* thin out token handling

* UI shows but assignRole fails

* working except for role removal

* Update manage-api.mjs

* disabled to prevent actions

The assignRole works, but removeRole does not. This prevents the Admins from making confusing data until it is repaired

---------

Co-authored-by: Bryan Haberberger <bryan.j.haberberger@slu.edu>

* enabled

* update packages to latest, except auth0 (#37)

* dev it

---------

Co-authored-by: cubap <cubap@slu.edu>